third_party/blink/web_tests/platform/generic/fast/xmlhttprequest/xmlhttprequest-bad-mimetype-expected.txt

This test checks for <rdar://problem/5303567> REGRESSION: XMLHttpRequest.responseXML returns NULL if response MIME type is not XML -- breaks Wikipedia widget

PASS: No exception.
readyState: 4 (number)
responseText: <?xml version="1.0" encoding="UTF-8"?>... (string)
responseXML: [object XMLDocument] (object)
status: 200 (number)
statusText: OK (string)
